Added validation & reload_db.sh

+++ trunk/trad4/bin/send_reload.sh
@@ -2,5 +2,12 @@
 
 PID=`ps -ef | grep $INSTANCE | grep -v grep | grep -v mysql | awk '{print $2}'`
 
-kill -10 $PID
+if [ $PID ]
+then
+    kill -10 $PID
+else
+    echo "$INSTANCE not found in ps. Are you sure $INSTANCE is running?"
+    exit 0
+fi
+
